New !
This is an inheritance project in C++ with virtual functions. The purpose of the project is to provide myself with functions to test. I wanted experience on what it is like to write tests cases for programs and thinking of different inputs to sub into each case. Also gathering an idea as to what test to run and what kind. In addition to this this is also a review of C++.